Calculate positioning relative to the appendTo element
The previous positioning did not take the `appendTo` option into
account. Consequently, the dropdown was not positioned properly in some
cases, for example inside of fixed Bootstrap dialogs. Appending to the
body breaks there when the dialog opens after the user scrolled the page
a bit.

With this patch, the dropdown can be appended to the dialog instead of
the body. The dropdown itself then does not need `position: fixed` and
always appears in the right spot.

Fixes the issue in the comments of #125