Added the ability to insert HTML into a "contenteditable" field
These changes allow the "replace" function of the strategy to return
HTML for both the pre and the post parts. The HTML is then inserted
correctly and the cursor positioned in between them. With this, it is
for example possible to directly insert smiley <img> etc.. This should
also resolve issue #105.