FIX: avoid dropdown bumping into right edge of window
If the autocomplete dropdown is posititioned too far right, the
container div for the dropdown gets clipped at the window width,
forcing its contained elements to wrap onto the next line.

The fix is to move farther left in this case; ideally we'd just
position at (left = window.width - dropdown.width); however AFAIK
there's no way to get the true width of the dropdown, only the clipped
width, because by the time you can ask the browser its width it's
already been clipped. So instead we employ a heuristic: if it's too
close to the right edge of the window (in my case, with a browser
window 1280px wide and a vertical scrollbar showing, it seems to get
clipped at 1264px, so "too close" must be > 16px; I set it to 30 here)
then we assume we got clipped, we move somewhat left which gives us
room to grow, and we see if we're still clipped, repeating as
necessary.