CKEditor Adapter - Fix Space Problem by &nbsp; + Delayed Search
Problem was, that spaces after the completion were "Eaten" by CKEditor.

Don't worry, the &nbsp; is replaced by a normal space on the next non-space anyways.

Search was delayed as well, i.e. the last key was not used in the search, that is why
we listen again to the native event.