Fix dropdown showing out of viewport
When 'top' is passed into options and you have a document height that's larger than the viewport, the dropdown can show outside of the screen because the parent height can be 4400px (see a Discourse site like [Cloud9 community](https://community.c9.io)). This takes the min of the window height and the parent height so it ensures the dropdown will show in the viewport.