Fix #299: ignore search callback after dropdown deactivate
If the dropdown is shown, but it then invokes the search callback, and
the search callback takes a while, and before it returns we hide the
dropdown, and then we call the search callback with the dropdown
hidden, jquery-textcomplete tries to show the dropdown again, but the
element for the dropdown is still hidden when we try to position
it. jquery-textcomplete asks jquery for the bounding rect of the
element, jquery asks the browser/DOM, Chrome returns a ClientRect
object that isn't literally frozen but acts like it, jquery notices it
has width and height 0 because it's hidden so it returns the
frozen-ish browser object passthrough instead of creating a writable
object, jquery-textcomplete tries to write to it, and boom. (And even
if we avoid the write to the frozen object, the completer's internal
state is messed up and the next time we try to show the dropdown, it
won't happen.)

We do better to prevent jquery-textcomplete from trying to show the
dropdown in response to a late-arriving search completion, by which I
mean, if the user cancels the dropdown while the search is in
progress, then when the search completes, we should just pretend it
didn't. To do this, I implemented a generation counter scheme, where
each time we start a new search, we increment the counter (stored in
the dropdown object, not the textcomplete object, because the
dropdown's deactivate method also needs to see it), and each time we
cancel the menu, we set a separate cancelled counter to the last
search counter, and each time a search completes, we see if the
cancelled counter is as high as the current search counter (meaning
the current search got cancelled) -- if so, we don't try to show the
dropdown again for this search.

Tested via Chrome's slow network simulator ("good 2G" preset).