Fix #299: avoid writing to browser ClientRect object
If we trigger textcomplete at the beginning of a line in a
contenteditable element in webkit-based browsers (Chrome and Safari),
then backspace over the trigger character before the search completes,
then the leaf element containing the cursor, where we create the
placeholder <span> element, is empty except for our placeholder <span>
element.

Then when we call $node.offset() to get its client rect, the browser
returns a ClientRect object with 0 height and width (fair enough --
the element is empty), and jQuery treats that as meaning the element
is hidden, and retuns the browser's own ClientRect object instead of a
new object with the offset properties. The browser's own ClientRect
object is not writable, and throws exceptions if you try to write to
it, which jquery-textcomplete does.

I worked around this by detecting the width=0 case and cloning the
object (just the properties we need) to a new mutable object so we
don't write to the one the browser and jQuery gave us.

Tested via Chrome's slow network simulator ("good 2G" preset).