Don't let contextFunc influence other strategies
This fixes issue #301.
Using contextFunc in one strategy should have no effect on other strategies.
But before this change, the contextFunc would replace the text passed to the remaining strategies also.
Instead now only the current strategy uses the context directly (or the original text if the strategy uses no context function). The text is not changed anymore.