On IE: Correct alignment of jquery-textcomplete-dropdown when scrolling down in iframe
On IE: When the jquery-textcomplete was placed in an iframe (e.g. ckeditor) the jquery-textcomplete dropdown was not aligned correctly when the user scrolled down in the textarea.

This was caused by the fact that scrollTop (of the body-element inside the iframe) is used to calculate the correct position of the dropdown. However on IE scrollTop of body is always zero. This a known issue: https://stackoverflow.com/questions/2717252/document-body-scrolltop-is-always-0-in-ie-even-when-scrolling

This was solved by getting the scrollTop of the html-element inside the iframe, which works for all browsers, including IE.